# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SabiScore API",
    "description": "Football Prediction API with advanced analytics and odds comparison",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/matches/upcoming": {
      "get": {
        "tags": [
          "matches"
        ],
        "summary": "Get Upcoming Matches",
        "description": "Get upcoming matches for predictions\n\nReturns matches scheduled within the next N days, optionally filtered by league.\nResults are cached for 5 minutes to optimize performance.",
        "operationId": "matches-get_upcoming_matches",
        "parameters": [
          {
            "name": "league",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by league (epl, bundesliga, laliga, etc)",
              "title": "League"
            },
            "description": "Filter by league (epl, bundesliga, laliga, etc)"
          },
          {
            "name": "days_ahead",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "description": "Number of days to look ahead",
              "default": 7,
              "title": "Days Ahead"
            },
            "description": "Number of days to look ahead"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of matches to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Maximum number of matches to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/matches/{match_id}": {
      "get": {
        "tags": [
          "matches"
        ],
        "summary": "Get Match Detail",
        "description": "Get detailed information about a specific match\n\nIncludes team form, head-to-head history, odds movement, and prediction metadata.",
        "operationId": "matches-get_match_detail",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Match Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/matches/league/{league_name}": {
      "get": {
        "tags": [
          "matches"
        ],
        "summary": "Get Matches By League",
        "description": "Get all matches for a specific league\n\nUseful for league-specific analysis and historical performance tracking.",
        "operationId": "matches-get_matches_by_league",
        "parameters": [
          {
            "name": "league_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "League Name"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status (scheduled, live, finished)",
              "title": "Status"
            },
            "description": "Filter by status (scheduled, live, finished)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/odds/match/{match_id}": {
      "get": {
        "tags": [
          "odds"
        ],
        "summary": "Get Match Odds",
        "description": "Get odds history for a specific match\n\nReturns timestamped odds snapshots showing market movement over time.\nUseful for identifying line movement and CLV opportunities.",
        "operationId": "odds-get_match_odds",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Match Id"
            }
          },
          {
            "name": "bookmaker",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by bookmaker",
              "title": "Bookmaker"
            },
            "description": "Filter by bookmaker"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of snapshots",
              "default": 10,
              "title": "Limit"
            },
            "description": "Maximum number of snapshots"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OddsResponse"
                  },
                  "title": "Response Odds-Get Match Odds"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/matches/search": {
      "get": {
        "summary": "Search Matches",
        "description": "Search for matches with filtering, pagination, and caching.",
        "operationId": "search_matches",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "description": "Search query for teams or matches",
              "title": "Q"
            },
            "description": "Search query for teams or matches"
          },
          {
            "name": "league",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by league",
              "title": "League"
            },
            "description": "Filter by league"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Maximum results to return",
              "default": 10,
              "title": "Limit"
            },
            "description": "Maximum results to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MatchSearchResponse"
                  },
                  "title": "Response Search Matches"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MatchDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "home_team": {
            "type": "string",
            "title": "Home Team"
          },
          "away_team": {
            "type": "string",
            "title": "Away Team"
          },
          "league": {
            "type": "string",
            "title": "League"
          },
          "match_date": {
            "type": "string",
            "format": "date-time",
            "title": "Match Date"
          },
          "venue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "scheduled"
          },
          "odds": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Odds"
          },
          "referee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referee"
          },
          "season": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season"
          },
          "round_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Round Number"
          }
        },
        "type": "object",
        "required": [
          "id",
          "home_team",
          "away_team",
          "league",
          "match_date"
        ],
        "title": "MatchDetailResponse",
        "description": "Lightweight match detail payload used by public endpoints."
      },
      "MatchListResponse": {
        "properties": {
          "matches": {
            "items": {
              "$ref": "#/components/schemas/MatchSummary"
            },
            "type": "array",
            "title": "Matches"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "league_filter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "League Filter"
          },
          "date_range_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Range Days"
          }
        },
        "type": "object",
        "required": [
          "matches",
          "total"
        ],
        "title": "MatchListResponse",
        "description": "Container for a list of matches with simple summary items."
      },
...
      "MatchSearchResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "example": "1"
          },
          "home_team": {
            "type": "string",
            "title": "Home Team",
            "example": "Manchester City"
          },
          "away_team": {
            "type": "string",
            "title": "Away Team",
            "example": "Liverpool"
          },
          "league": {
            "type": "string",
            "title": "League",
            "example": "EPL"
          },
          "match_date": {
            "type": "string",
            "title": "Match Date",
            "example": "2024-10-26T15:00:00Z"
          },
          "venue": {
            "type": "string",
            "title": "Venue",
            "example": "Etihad Stadium"
          }
        },
        "type": "object",
        "required": [
          "id",
          "home_team",
          "away_team",
          "league",
          "match_date",
          "venue"
        ],
        "title": "MatchSearchResponse"
      },
      "MatchSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "home_team": {
            "type": "string",
            "title": "Home Team"
          },
          "away_team": {
            "type": "string",
            "title": "Away Team"
          },
          "league": {
            "type": "string",
            "title": "League"
          },
          "match_date": {
            "type": "string",
            "format": "date-time",
            "title": "Match Date"
          },
          "venue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "scheduled"
          },
          "has_odds": {
            "type": "boolean",
            "title": "Has Odds",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "home_team",
          "away_team",
          "league",
          "match_date"
        ],
        "title": "MatchSummary",
        "description": "Lean response tailored for listings and search results."
      },
...
      "OddsResponse": {
        "properties": {
          "match_id": {
            "type": "string",
            "title": "Match Id",
            "description": "Associated match identifier"
          },
          "bookmaker": {
            "type": "string",
            "title": "Bookmaker",
            "description": "Bookmaker name"
          },
          "home_win": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 1.01
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Win"
          },
          "draw": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 1.01
              },
              {
                "type": "null"
              }
            ],
            "title": "Draw"
          },
          "away_win": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 1.01
              },
              {
                "type": "null"
              }
            ],
            "title": "Away Win"
          },
          "over_under": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 1.01
              },
              {
                "type": "null"
              }
            ],
            "title": "Over Under"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "match_id",
          "bookmaker"
        ],
        "title": "OddsResponse",
        "description": "API response model for odds data."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}